합본
합본

+++ Server/Server/ChatManager.cpp
@@ -1,3 +1,4 @@
+#include "stdafx.h"
 #include "LobbyManager.h"
 #include "ChatManager.h"
 #include "C_ClientInfo.h"
@@ -27,7 +28,7 @@ void ChatManager::End()
 void ChatManager::PackPacket(char* _setptr, TCHAR* _str1, int& _size)
 {
 	char* ptr = _setptr;
-	int strsize1 = _tcslen(_str1) * sizeof(TCHAR);
+	int strsize1 = (int)_tcslen(_str1) * sizeof(TCHAR);
 	_size = 0;
 
 	// 문자열 길이
